pyproject.toml
[project]
name = "sld-interpreter"
version = "0.1.0"
description = "Production-grade Single Line Diagram interpretation system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "opencv-python-headless>=4.9.0",
    "Pillow>=10.2.0",
    "numpy>=1.26.0",
    "scikit-image>=0.22.0",
    "pytesseract>=0.3.10",
    " paddleocr>=2.7.0",
    "ultralytics>=8.1.0",
    "networkx>=3.2.0",
    "shapely>=2.0.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
    "celery>=5.3.0",
    "boto3>=1.34.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "httpx>=0.26.0",
    "jinja2>=3.1.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
